+++ src/app/lhc-group-horizontal/lhc-group-horizontal.component.ts
@@ -2,9 +2,10 @@ import { Component, OnInit, OnChanges, Input } from '@angular/core';
 import { LhcDataService} from '../../lib/lhc-data.service';
 
 @Component({
-  selector: 'lhc-group-horizontal',
-  templateUrl: './lhc-group-horizontal.component.html',
-  styleUrls: ['./lhc-group-horizontal.component.css']
+    selector: 'lhc-group-horizontal',
+    templateUrl: './lhc-group-horizontal.component.html',
+    styleUrls: ['./lhc-group-horizontal.component.css'],
+    standalone: false
 })
 export class LhcGroupHorizontalComponent {