+++ src/app/lhc-item-choice/lhc-item-choice.component.ts
@@ -1,9 +1,10 @@
 import { Component, Input, OnInit, OnChanges } from '@angular/core';
 
 @Component({
-  selector: 'lhc-item-choice',
-  templateUrl: './lhc-item-choice.component.html',
-  styleUrls: ['./lhc-item-choice.component.css']
+    selector: 'lhc-item-choice',
+    templateUrl: './lhc-item-choice.component.html',
+    styleUrls: ['./lhc-item-choice.component.css'],
+    standalone: false
 })
 export class LhcItemChoiceComponent {