+++ src/fhir/R4R5-common/sdc-import.js
@@ -34,18 +34,17 @@ function addSDCImportFns(ns) {
       questionnaire.contained.forEach(function (vs) {
         if(vs.resourceType === 'ValueSet') {
           var answers = self.answersFromVS(vs);
-          if (!answers)
-            answers = []; // continuing with previous default; not sure if needed
-
-          // Support both id and url based lookup - we are only supporting our non-standard url approach
-          // for backward-compatibility with previous LForms versions. For more details on FHIR contained
-          // resource references, please see "http://hl7.org/fhir/references.html#canonical-fragments"
-          var lfVS = {answers: answers};
-          if(vs.id) {
-            answersVS['#' + vs.id] = lfVS;
-          }
-          if(vs.url) {
-            answersVS[vs.url] = lfVS;
+          if (answers) {
+            // Support both id and url based lookup - we are only supporting our non-standard url approach
+            // for backward-compatibility with previous LForms versions. For more details on FHIR contained
+            // resource references, please see "http://hl7.org/fhir/references.html#canonical-fragments"
+            var lfVS = {answers: answers};
+            if(vs.id) {
+              answersVS['#' + vs.id] = lfVS;
+            }
+            if(vs.url) {
+              answersVS[vs.url] = lfVS;
+            }
           }
         }
       });
@@ -175,11 +174,13 @@ function addSDCImportFns(ns) {
     else if (qItem.answerValueSet) {
       if (containedVS)
         var vs = containedVS[qItem.answerValueSet];
-      if(vs) { // contained
+      if(vs && vs.answers) { // contained ValueSet with an expansion
         lfItem.answers = vs.answers;
+        // To keep answerValueSet property during export.
+        lfItem._answerValueSet = qItem.answerValueSet;
       }
       else
-        lfItem.answerValueSet = qItem.answerValueSet; // a URI for a ValueSet
+        lfItem.answerValueSet = qItem.answerValueSet; // a URI for a ValueSet, or an ID for a contained ValueSet with no expansion
     }
   };