+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.11.0] 2025-02-10
+## [36.15.0] 2025-02-10
 ### Changed
 - The promise returned by LForms.Util.addFormToPage, when it fails, will now
   reject with an array of error messages instead of just a single error message.
@@ -11,6 +11,52 @@ This project follows [Semantic Versioning](http://semver.org/).
   This is potentially a breaking change, but the impact is unlikely to be
   harmful given the way error messages are typically used, so we did not change
   the major version.
+
+## [36.14.0] 2025-03-13
+### Added
+- Support for itemControl 'legal'.
+
+## [36.13.0] 2025-03-06
+### Added
+- Support for rendering-xhtml on externally loaded answerValueSet prefetch/radio/checkbox.
+
+## [36.12.0] 2025-03-06
+### Added
+- Loading answerValueSet for contained ValueSets without an expansion.
+### Fixed
+- Exported Questionnaire should have answerValueSet property but not populated 
+  answerOptions if they are from contained ValueSets.
+
+## [36.11.2] 2025-03-03
+### Fixed
+- Only the new externallyDefined URL should be exported.
+
+## [36.11.1] 2025-02-28
+### Changed
+- New URL for preferredTerminologyServer extension.
+- preferredTerminologyServer extension should work at root level as well as
+  item level.
+
+## [36.11.0] 2025-02-27
+### Changed
+- Upgraded angular to v19
+- Upgraded ng-zorro-antd to v19
+
+## [36.10.5] 2025-02-25
+### Fixed
+- Fixed a display issue for answers, where the display field might
+  not have data.
+
+## [36.10.4] 2025-02-24
+### Fixed
+- Fixed a bug that the repeating numbers are not displayed correctly 
+  on the repeating items when their data are loaded through FHIRPath
+  expressions.
+  
+## [36.10.3] 2025-02-07
+### Changed
+- Changed the meta.profile for STU3, R4 and R5 in the exported 
+  QuestionnaireResponse so that the standard FHIR profiles are used.
 
 ## [36.10.2] 2025-02-03
 ### Fixed